백준2748번 int -> long

+++ src/baekjoon/p10870/Main.java
@@ -12,7 +12,7 @@ public class Main {
             System.out.println(0);
             return;
         }
-        int[] fib = new int[n+1];
+        long[] fib = new long[n+1];
         fib[1] = 1;
         for(int i=2;i<=n;i++){
             fib[i] = fib[i - 1] + fib[i - 2];
